refactor: split storage runtime view modules
